# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Flowable SaaS PoC API
  version: 0.1.0
paths:
...
  /api/tasks/{taskId}/returnTo:
    post:
      summary: Return task to a specified historic activity
      parameters:
        - name: taskId
          in: path
          required: true
          schema: { type: string }
        - name: targetActivityId
          in: query
          required: true
          schema: { type: string }
        - name: reason
          in: query
          required: true
          schema: { type: string }
      responses:
        '200': { description: returned }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT